refactor(evm): remove tss as much as possible
