+++ client/docs/static/swagger/swagger.yaml
@@ -2219,14 +2219,6 @@ paths:
                 format: byte
               name:
                 type: string
-              key_type:
-                type: string
-                enum:
-                  - KEY_TYPE_UNSPECIFIED
-                  - KEY_TYPE_NONE
-                  - KEY_TYPE_THRESHOLD
-                  - KEY_TYPE_MULTISIG
-                default: KEY_TYPE_UNSPECIFIED
               params:
                 type: string
                 format: byte
@@ -42007,14 +41999,6 @@ definitions:
         format: byte
       name:
         type: string
-      key_type:
-        type: string
-        enum:
-          - KEY_TYPE_UNSPECIFIED
-          - KEY_TYPE_NONE
-          - KEY_TYPE_THRESHOLD
-          - KEY_TYPE_MULTISIG
-        default: KEY_TYPE_UNSPECIFIED
       params:
         type: string
         format: byte
