+++ x/evm/types/msg_add_chain.go
@@ -8,16 +8,14 @@ import (
 
 	"github.com/axelarnetwork/axelar-core/utils"
 	nexus "github.com/axelarnetwork/axelar-core/x/nexus/exported"
-	tss "github.com/axelarnetwork/axelar-core/x/tss/exported"
 )
 
 // NewAddChainRequest is the constructor for NewAddChainRequest
-func NewAddChainRequest(sender sdk.AccAddress, name string, keyType tss.KeyType, params Params) *AddChainRequest {
+func NewAddChainRequest(sender sdk.AccAddress, name string, params Params) *AddChainRequest {
 	return &AddChainRequest{
-		Sender:  sender,
-		Name:    nexus.ChainName(utils.NormalizeString(name)),
-		KeyType: keyType,
-		Params:  params,
+		Sender: sender,
+		Name:   nexus.ChainName(utils.NormalizeString(name)),
+		Params: params,
 	}
 }
 
@@ -40,7 +38,6 @@ func (m AddChainRequest) ValidateBasic() error {
 	chain := nexus.Chain{
 		Name:                  m.Name,
 		SupportsForeignAssets: true,
-		KeyType:               m.KeyType,
 		Module:                ModuleName,
 	}
