+++ x/evm/types/msg_create_transfer_operatorship.go
@@ -5,16 +5,16 @@ import (
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 
 	"github.com/axelarnetwork/axelar-core/utils"
+	multisig "github.com/axelarnetwork/axelar-core/x/multisig/exported"
 	nexus "github.com/axelarnetwork/axelar-core/x/nexus/exported"
-	tss "github.com/axelarnetwork/axelar-core/x/tss/exported"
 )
 
 // NewCreateTransferOperatorshipRequest creates a message of type CreateTransferOperatorshipRequest
-func NewCreateTransferOperatorshipRequest(sender sdk.AccAddress, chain string, keyID string) *CreateTransferOperatorshipRequest {
+func NewCreateTransferOperatorshipRequest(sender sdk.AccAddress, chain string, keyID multisig.KeyID) *CreateTransferOperatorshipRequest {
 	return &CreateTransferOperatorshipRequest{
 		Sender: sender,
 		Chain:  nexus.ChainName(utils.NormalizeString(chain)),
-		KeyID:  tss.KeyID(keyID),
+		KeyID:  keyID,
 	}
 }
 
@@ -38,7 +38,7 @@ func (m CreateTransferOperatorshipRequest) ValidateBasic() error {
 		return sdkerrors.Wrap(err, "invalid chain")
 	}
 
-	if err := m.KeyID.Validate(); err != nil {
+	if err := m.KeyID.ValidateBasic(); err != nil {
 		return err
 	}
