+++ x/evm/types/msg_create_transfer_ownership.go
@@ -5,13 +5,17 @@ import (
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 
 	"github.com/axelarnetwork/axelar-core/utils"
+	multisig "github.com/axelarnetwork/axelar-core/x/multisig/exported"
 	nexus "github.com/axelarnetwork/axelar-core/x/nexus/exported"
-	tss "github.com/axelarnetwork/axelar-core/x/tss/exported"
 )
 
 // NewCreateTransferOwnershipRequest is the constructor for CreateTransferOwnershipRequest
-func NewCreateTransferOwnershipRequest(sender sdk.AccAddress, chain string, keyID string) *CreateTransferOwnershipRequest {
-	return &CreateTransferOwnershipRequest{Sender: sender, Chain: nexus.ChainName(utils.NormalizeString(chain)), KeyID: tss.KeyID(keyID)}
+func NewCreateTransferOwnershipRequest(sender sdk.AccAddress, chain string, keyID multisig.KeyID) *CreateTransferOwnershipRequest {
+	return &CreateTransferOwnershipRequest{
+		Sender: sender,
+		Chain:  nexus.ChainName(utils.NormalizeString(chain)),
+		KeyID:  keyID,
+	}
 }
 
 // Route implements sdk.Msg
@@ -44,7 +48,7 @@ func (m CreateTransferOwnershipRequest) ValidateBasic() error {
 		return sdkerrors.Wrap(err, "invalid chain")
 	}
 
-	if err := m.KeyID.Validate(); err != nil {
+	if err := m.KeyID.ValidateBasic(); err != nil {
 		return err
 	}
